use Server.Shutdown method to avoid overhead
The HTTP Server has support for graceful shutdown in Go 1.8
https://beta.golang.org/doc/go1.8#http_shutdown
go-gracedown is now just a wrapper of the net/http package to maintain interface compatibility.
