+++ .travis.yml
@@ -4,4 +4,5 @@ go:
   - "1.4"
   - "1.5"
   - "1.6"
+  - "1.7"
   - tip
